feat(compiler-project): :sparkles: simple visitor
made a first simple pretty printer to rewrite the input code.
This is part of the minimal compiler for the project.
This is still in progress. I need to implement more in each parts.
(lexer, parser, visitors)
The next steps will be to add more supported Dart syntax and a simple
code generator. (visitor for converting the AST to a UML diagram).